opt in to at least once sampling
